requirements.txt
Django~=5.0.6
djangorestframework~=3.15.1
psycopg2-binary~=2.9.9
python-environ~=0.4.54
djangorestframework-simplejwt~=5.3.1
dadata~=21.10.1
redis~=5.0.6